items/backend/requirements.txt
asgiref
Django
psycopg
djangorestframework
djangorestframework-simplejwt
sqlparse
django-auth-ldap
django-environ
django-cors-headers